# ClassValueCompat to support systems without java.lang.ClassValue