… src/library/scala/runtime/ModuleSerializationProxy.java