The goal of this commit is to enable systems without java.lang.ClassValue to fall back to other method which does not utilize java.lang.ClassValue while keeping its current behavior as how it is for systems with java.lang.ClassValue. Use of java.lang.ClassValue seems mainly for caching as https://github.com/scala/scala/pull/7879 introduced, so I ran a related benchmark from one of the commits and got following results

```
Baseline

[info] Benchmark                                             Mode  Cnt   Score    Error   Units
[info] ClassTagBenchmark.lookupClassTag                      avgt   20  40.328 ±  0.496   ns/op
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate       avgt   20  ≈ 10⁻⁴           MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate.norm  avgt   20  ≈ 10⁻⁵             B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.count            avgt   20     ≈ 0           counts
```

```
ClassValueCompat - java.lang.ClassValue exists

[info] Benchmark                                             Mode  Cnt   Score    Error   Units
[info] ClassTagBenchmark.lookupClassTag                      avgt   20  42.834 ±  0.231   ns/op
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate       avgt   20  ≈ 10⁻⁴           MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate.norm  avgt   20  ≈ 10⁻⁵             B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.count            avgt   20     ≈ 0           counts
```

For cases where systems lacking java.lang.ClassValue, I earned the following result which seems obviously slower than the above two but still it enables such systems to actually utilize classes such as ClassTag instead of throwing ClassNotFoundException

```
ClassValueCompat - java.lang.ClassValue does not exist

[info] Benchmark                                                      Mode  Cnt     Score    Error   Units
[info] ClassTagBenchmark.lookupClassTag                               avgt   20    79.452 ±  0.294   ns/op
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate                avgt   20  3838.471 ± 15.265  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.alloc.rate.norm           avgt   20   336.000 ±  0.001    B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.G1_Eden_Space       avgt   20  3843.122 ± 24.785  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.G1_Eden_Space.norm  avgt   20   336.408 ±  1.792    B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.G1_Old_Gen          avgt   20     0.003 ±  0.001  MB/sec
[info] ClassTagBenchmark.lookupClassTag:·gc.churn.G1_Old_Gen.norm     avgt   20    ≈ 10⁻⁴             B/op
[info] ClassTagBenchmark.lookupClassTag:·gc.count                     avgt   20  1525.000           counts
[info] ClassTagBenchmark.lookupClassTag:·gc.time                      avgt   20   655.000               ms
```

This commit combined with https://github.com/scala/scala/pull/9739 will open Scala 2.13.x to wider range of systems including Android as discussed and demonstrated at https://users.scala-lang.org/t/scala-2-13-on-android/7235/24