@smarter 2021-09-07
Nice! FWIW, I've also opened an issue against Android about the lack of ClassValue, but it's not clear to me if they're interested in fixing it: https://issuetracker.google.com/issues/196063118 (maybe someone with some twitter clout could try bringing attention to this?)