@SethTisue 2021-10-12
on JDK 17 we're seeing (at https://app.travis-ci.com/github/scala/scala/jobs/542786102):

```
[error] /home/travis/build/scala/scala/src/library/scala/runtime/ModuleSerializationProxy.scala:23:23: class AccessController in package security is deprecated
[error]         java.security.AccessController.doPrivileged((() => cls.getField("MODULE$").get(null)): PrivilegedExceptionAction[Object])
[error]                       ^
```

PR with fix: #9785