@fiatjaf 2022-07-26
I have a question related to this PR, so I thought I would just beg for attention here. Let me know if this is inappropriate: https://stackoverflow.com/questions/73042201/why-a-scala-2-13-app-is-trying-to-use-javaclassvalue-on-android-if-it-is-not-a